Upgrade html-to-text
The html-to-text package has a vulnerability and needs upgrading.

The updated library seems to have better performance, be a
slightly smaller dependency tree download (owing to the removal of
underscore) and is more configurable (not that we need at the moment).
I tooks advantage of the library's new compile function, which should
improve performance.