Don't transform bodyHTML to plain text
Until this change, article and blog post bodyHTML would be transformed
into plain text if no plain bodyText was present. bodyText is almost always
available when the content is retrieved from Next ElasticSearch.
However, this can be very slow to perform for long articles. If Next
ElasticSearch suddenly didn't have this data for some reason, every
article page request would have to do this transformation.

It's an optional property, so probably not worth risking the
application resources on it if this were to happen [^1].

The only Articles without bodyText in Next Elastic seem to be
decommissioned Alphaville content that actually redirects to the Alphaville
stream page.

For Live Blog summaries the functionality remains the same: we're still
tranforming that HTML to plain text because plainText isn't present in
Next ElasticSearch (perhaps it ought to be?).

[^1]: https://developers.google.com/search/docs/advanced/structured-data/article#non-amp-sd